update for half

+++ projects/clr/hipamd/include/hip/amd_detail/amd_hip_fp8.h
@@ -1126,7 +1126,7 @@ struct __hip_fp8_e8m0 {
 
   __hip_fp8_e8m0() = default;
   __FP8_HOST_DEVICE__ inline explicit __hip_fp8_e8m0(const __half f) {
-    __x = __hip_cvt_float_to_e8m0(f, __default_saturation, __default_interpret);
+    __x = __hip_cvt_float_to_e8m0(__half2float(f), __default_saturation, __default_interpret);
   }
   __FP8_HOST_DEVICE__ inline explicit __hip_fp8_e8m0(const __hip_bfloat16 f) {
     __x = __hip_cvt_bfloat16raw_to_e8m0(f, __default_saturation, __default_interpret);
